Move module `environment` in subfolder `src/bed_bpp_env`
