+++ src/bed_bpp_env/environment/height_map.py
